add BlockDCT21, DistortionEstimation23
jzlk_DCTII_v2: function, performs DCTII on input
jzlk_DCTII: function, performs DCTII on input using fft (only 1D)
jzlk_MatQuant: function, performs quantization of matrix

+++ Project 1/project_1/newcode/jzlk_wienerFilterWithoutTapering.m
@@ -15,7 +15,7 @@ H = fft2(h, size(g,1), size(g,2));
 
 % The exact value for K is given by K = sigma2/var(f). Approximate var(f)
 % by var(g)/2
-K = 2* sigma2 / var(g(:));
+K = 2* sigma2 / var(double(g(:)));
 
 % Calculate Wiener Transfer Function
 H2 = abs(H).^2;
